+++ test-fixtures/config/vscodeconfig/test.result.js
@@ -1,0 +1,16 @@
+function foo() {
+  const foo = [
+    "aaaaaaaaa",
+    "bbbbbb",
+    "c",
+    "a",
+    "b",
+    "c",
+    "a",
+    "b",
+    "c",
+    "a",
+    "b",
+    "c",
+  ];
+}