+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "prettier-vscode" extension will be documented in thi
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
+## [3.8.0]
+
+- Fixed issue where VS Code and local config where merged. If local config is present, only it will be used. #1074
+
 ## [3.7.0]
 
 - Removed deprecation message from `requireConfig` (Was added by mistake). #1056